Move and rename services
