+++ packages/backend/src/app/Links/repository/LinkRepositoryImpl.ts
@@ -1,10 +1,10 @@
 import { Inject, Service } from 'typedi';
 import { Repository } from 'typeorm';
 import NotFoundError from '../../../errors/NotFoundError';
-import ILogger, { Logger } from '../../../infrastructure/Logger/Logger';
-import LoggerImpl from '../../../infrastructure/Logger/LoggerImpl';
-import ITypeOrmService from '../../../services/TypeOrmService';
-import TypeOrmServiceImpl from '../../../services/TypeOrmServiceImpl';
+import ILoggerService, { Logger } from '../../../services/LoggerService/LoggerService';
+import LoggerServiceImpl from '../../../services/LoggerService/LoggerServiceImpl';
+import ITypeOrmService from '../../../services/TypeOrmService/TypeOrmService';
+import TypeOrmServiceImpl from '../../../services/TypeOrmService/TypeOrmServiceImpl';
 import ITagService from '../../Tags/service/TagService';
 import TagServiceImpl from '../../Tags/service/TagServiceImpl';
 import User from '../../Users/User';
@@ -19,7 +19,7 @@ export default class LinkRepositoryImpl implements ILinkRepository {
   private readonly repository: Repository<LinkModel>;
 
   constructor(
-    @Inject(() => LoggerImpl) private readonly logger: ILogger,
+    @Inject(() => LoggerServiceImpl) private readonly logger: ILoggerService,
     @Inject(() => TypeOrmServiceImpl) private readonly typeOrm: ITypeOrmService,
     @Inject(() => TagServiceImpl) private readonly tagService: ITagService,
   ) {
