+++ packages/backend/src/app/Tags/repository/TagRepositoryImpl.ts
@@ -1,9 +1,9 @@
 import { Inject, Service } from 'typedi';
 import { Repository } from 'typeorm';
-import ILogger, { Logger } from '../../../infrastructure/Logger/Logger';
-import LoggerImpl from '../../../infrastructure/Logger/LoggerImpl';
-import ITypeOrmService from '../../../services/TypeOrmService';
-import TypeOrmServiceImpl from '../../../services/TypeOrmServiceImpl';
+import ILoggerService, { Logger } from '../../../services/LoggerService/LoggerService';
+import LoggerServiceImpl from '../../../services/LoggerService/LoggerServiceImpl';
+import ITypeOrmService from '../../../services/TypeOrmService/TypeOrmService';
+import TypeOrmServiceImpl from '../../../services/TypeOrmService/TypeOrmServiceImpl';
 import User from '../../Users/User';
 import Tag from '../Tag';
 import TagModel from './TagModel.entity';
@@ -16,7 +16,7 @@ export default class TagRepositoryImpl implements ITagRepository {
   private readonly repository: Repository<TagModel>;
 
   constructor(
-    @Inject(() => LoggerImpl) private readonly logger: ILogger,
+    @Inject(() => LoggerServiceImpl) private readonly logger: ILoggerService,
     @Inject(() => TypeOrmServiceImpl) private readonly typeOrm: ITypeOrmService,
   ) {
     this.log = this.logger.child('TagRepository');
