+++ packages/backend/src/app/Users/service/UserServiceImpl.ts
@@ -1,6 +1,6 @@
 import { Inject, Service } from 'typedi';
-import ILogger, { Logger } from '../../../infrastructure/Logger/Logger';
-import LoggerImpl from '../../../infrastructure/Logger/LoggerImpl';
+import ILoggerService, { Logger } from '../../../services/LoggerService/LoggerService';
+import LoggerServiceImpl from '../../../services/LoggerService/LoggerServiceImpl';
 import IUserRepository from '../repository/UserRepository';
 import UserRepositoryImpl from '../repository/UserRepositoryImpl';
 import IUserService from './UserService';
@@ -10,7 +10,7 @@ export default class UserServiceImpl implements IUserService {
   private readonly log: Logger;
 
   constructor(
-    @Inject(() => LoggerImpl) private readonly logger: ILogger,
+    @Inject(() => LoggerServiceImpl) private readonly logger: ILoggerService,
     @Inject(() => UserRepositoryImpl) private readonly userRepository: IUserRepository,
   ) {
     this.log = this.logger.child('UserService');
